+++ include/mruby_integration/audio/sound.hpp
@@ -1,6 +1,0 @@
-#pragma once
-
-#include "mruby.h"
-
-void
-append_audio_sound(mrb_state*);